+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.4] - 2020-01-11
+
+### Added
+- query object can use `range` to set specific time windows
+- appoptics added modulo and power binary operations
+- query class has 100% code coverage
+
 ## [0.1.3] - 2019-12-07
 ### Added
 - improved label handling/reduction